fix(mail): BigInt IMAP UIDs (#10) — no signed-32-bit overflow
IMAP UIDs are unsigned 32-bit and can exceed 2^31 on long-lived busy mailboxes.
EmailQueue.lastSeenUid and InboundDelivery.uid are now BigInt (were signed INTEGER).
The poll computes in Number (UIDs are well within 2^53) and stores back as BigInt; the
cursor CAS compares/writes BigInt. 265 passed; tsc + lint clean.

+++ apps/api/prisma/migrations/20260717000000_inbound_delivery_ledger/migration.sql
@@ -7,7 +7,7 @@ CREATE TYPE "InboundDeliveryState" AS ENUM ('ACCEPTED', 'PROCESSING', 'PROCESSED
 CREATE TYPE "EmailQueueSyncState" AS ENUM ('OK', 'NEEDS_RECONCILIATION');
 
 -- Per-queue durable IMAP cursor + sync health (replaces Setting `imap/lastSeenUid:<id>`).
-ALTER TABLE "EmailQueue" ADD COLUMN "lastSeenUid" INTEGER NOT NULL DEFAULT 0;
+ALTER TABLE "EmailQueue" ADD COLUMN "lastSeenUid" BIGINT NOT NULL DEFAULT 0;
 ALTER TABLE "EmailQueue" ADD COLUMN "uidValidity" BIGINT;
 ALTER TABLE "EmailQueue" ADD COLUMN "syncState" "EmailQueueSyncState" NOT NULL DEFAULT 'OK';
 ALTER TABLE "EmailQueue" ADD COLUMN "lastError" TEXT;
@@ -20,7 +20,7 @@ CREATE TABLE "InboundDelivery" (
     "queueId" INTEGER,
     "transportKey" TEXT NOT NULL,
     "uidValidity" BIGINT,
-    "uid" INTEGER,
+    "uid" BIGINT,
     "externalId" TEXT,
     "messageId" TEXT,
     "contentHash" TEXT NOT NULL,
